feat(rust): support sql api
