+++ rust/lance/src/dataset/sql.rs
@@ -3,12 +3,32 @@
 
 use crate::datafusion::LanceTableProvider;
 use crate::dataset::SqlOptions;
+use crate::Dataset;
 use arrow_array::RecordBatch;
 use datafusion::execution::SendableRecordBatchStream;
 use datafusion::prelude::SessionContext;
 use lance_core::Error;
 use snafu::location;
 use std::sync::Arc;
+
+/// Customize the params of dataset's sql API.
+#[derive(Clone, Debug)]
+pub struct SqlOptions {
+    /// the dataset to run the SQL query
+    dataset: Option<Dataset>,
+
+    /// the SQL query to run
+    sql: String,
+
+    /// the name of the table to register in the datafusion context
+    table_name: String,
+
+    /// if true, the query result will include the internal row id
+    row_id: bool,
+
+    /// if true, the query result will include the internal row address
+    row_addr: bool,
+}
 
 impl SqlOptions {
     pub fn table_name(mut self, table_name: &str) -> Self {
